requirements.txt
numpy
opencv-python-headless
streamlit
mediapipe
streamlit-webrtc
librosa
sounddevice
scipy
parselmouth-praat